+++ CHANGELOG.rst
@@ -20,6 +20,8 @@ The format is based on `Keep a Changelog`_.
 
 Added
 -----
+- Introduced new options for handling credentials (qiskitrc file, environment
+  variables) and automatic registration. (#547)
 
 Changed
 -------